refactor(src): 优化 `main/` 和 `plugins/` 目录 & 更新导入
- 将 `main.ts` 从 `plugins/` 目录移动到上级 `src/目录
- 更新相关文件的导入路径
- 新增 `plugins/index.ts` 文件，统一导出插件功能
- 删除 `main/main.ts` 文件中的冗余代码
- 更新文件中的导入路径
