+++ src/plugins/index.ts
@@ -1,0 +1,11 @@
+export * from './help'
+export * from './chatSpawn'
+export * from './command'
+export * from './breakBlock'
+export * from './youAreMine'
+export * from './task'
+export * from './gui'
+export * from './autoFishing'
+export * from './killedBySimPlayer'
+export * from './setting'
+export * from './showCommandsList'
